refactor(docs): rename timing utility
